# frontend/src/lib/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "InvoiceFlow API",
    "description": "E-invoice processing platform with AI extraction and global compliance (PEPPOL, ZATCA, Brazil NFe, India IRP, Mexico CFDI, CTC reporting, and post-audit archival).",
    "version": "0.2.0"
  },
  "servers": [
    { "url": "http://localhost:8000", "description": "Local development" },
    { "url": "/api/v1", "description": "Production (Nginx proxy)" }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from POST /api/v1/auth/login"
      }
    },
    "schemas": {
      "InvoiceStatus": {
        "type": "string",
        "enum": ["received", "processing", "extracted", "reviewing", "approved", "compliant", "transmitted", "failed", "rejected"]
      },
      "ComplianceModel": {
        "type": "string",
        "enum": ["peppol", "clearance", "ctc", "post_audit"]
      },
      "ExtractionMethod": {
        "type": "string",
        "enum": ["ocr", "llm", "xml", "manual"]
      },
      "IngestionSource": {
        "type": "string",
        "enum": ["email", "ftp", "sftp", "api", "webhook", "manual"]
      },
      "InvoiceLine": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "invoice_id": { "type": "string", "format": "uuid" },
          "line_number": { "type": "integer" },
          "description": { "type": "string", "nullable": true },
          "quantity": { "type": "number" },
          "unit_price": { "type": "number" },
          "tax_rate": { "type": "number" },
          "tax_amount": { "type": "number" },
          "line_total": { "type": "number" },
          "item_code": { "type": "string", "nullable": true }
        },
        "required": ["id", "invoice_id", "line_number", "quantity", "unit_price", "tax_rate", "tax_amount", "line_total"]
      },
...
      "Invoice": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "tenant_id": { "type": "string", "format": "uuid" },
          "status": { "$ref": "#/components/schemas/InvoiceStatus" },
          "invoice_number": { "type": "string" },
          "vendor_name": { "type": "string" },
          "vendor_tax_id": { "type": "string", "nullable": true },
          "buyer_name": { "type": "string", "nullable": true },
          "buyer_tax_id": { "type": "string", "nullable": true },
          "invoice_date": { "type": "string", "format": "date-time", "nullable": true },
          "due_date": { "type": "string", "format": "date-time", "nullable": true },
          "currency": { "type": "string" },
          "subtotal": { "type": "number" },
          "tax_amount": { "type": "number" },
          "total_amount": { "type": "number" },
          "country_code": { "type": "string" },
          "compliance_model": { "$ref": "#/components/schemas/ComplianceModel" },
          "source": { "$ref": "#/components/schemas/IngestionSource" },
          "source_reference": { "type": "string", "nullable": true },
          "file_url": { "type": "string", "nullable": true },
          "original_filename": { "type": "string", "nullable": true },
          "ocr_confidence": { "type": "number", "nullable": true },
          "extraction_method": { "$ref": "#/components/schemas/ExtractionMethod" },
          "extracted_data": { "type": "object", "additionalProperties": true },
          "compliance_response": { "type": "object", "additionalProperties": true },
          "error_message": { "type": "string", "nullable": true },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" },
          "lines": { "type": "array", "items": { "$ref": "#/components/schemas/InvoiceLine" } },
          "audit_log": { "type": "array", "items": { "$ref": "#/components/schemas/AuditLog" }, "description": "Audit trail entries (included in detail view)" }
        },
        "required": ["id", "tenant_id", "status", "invoice_number", "vendor_name", "currency", "subtotal", "tax_amount", "total_amount", "country_code", "compliance_model", "source", "created_at", "updated_at"]
      },
...
      "InvoiceUpdate": {
        "type": "object",
        "properties": {
          "invoice_number": { "type": "string", "nullable": true },
          "vendor_name": { "type": "string", "nullable": true },
          "vendor_tax_id": { "type": "string", "nullable": true },
          "buyer_name": { "type": "string", "nullable": true },
          "buyer_tax_id": { "type": "string", "nullable": true },
          "invoice_date": { "type": "string", "format": "date-time", "nullable": true },
          "due_date": { "type": "string", "format": "date-time", "nullable": true },
          "currency": { "type": "string", "nullable": true },
          "status": { "$ref": "#/components/schemas/InvoiceStatus" }
        }
      },
...
      "AuditLog": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "tenant_id": { "type": "string", "format": "uuid" },
          "invoice_id": { "type": "string", "format": "uuid", "nullable": true },
          "action": { "type": "string" },
          "details": { "type": "object", "additionalProperties": true },
          "user_id": { "type": "string", "nullable": true },
          "timestamp": { "type": "string", "format": "date-time" }
        },
        "required": ["id", "tenant_id", "action", "timestamp"]
      },
...
      "FileUploadResponse": {
        "type": "object",
        "properties": {
          "invoice_id": { "type": "string", "format": "uuid", "nullable": true, "description": "Null when processing is async; use task_id for tracking" },
          "status": { "$ref": "#/components/schemas/InvoiceStatus" },
          "filename": { "type": "string" },
          "message": { "type": "string" }
        },
        "required": ["status", "filename", "message"]
      },
...
      "WebhookPayload": {
        "type": "object",
        "properties": {
          "filename": { "type": "string" },
          "file_content_b64": { "type": "string", "description": "Base64-encoded file content" },
          "mime_type": { "type": "string", "default": "application/pdf" },
          "source_reference": { "type": "string" },
          "metadata": { "type": "object", "additionalProperties": true }
        },
        "required": ["filename", "file_content_b64"]
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "detail": { "type": "string" }
        },
        "required": ["detail"]
      }
    }
  },
  "paths": {
...
    "/api/v1/invoices/{invoice_id}": {
      "get": {
        "tags": ["invoices"],
        "summary": "Get invoice by ID",
        "operationId": "getInvoice",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "name": "invoice_id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": {
          "200": { "description": "Invoice details", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Invoice" } } } },
          "404": { "description": "Invoice not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      },
      "put": {
        "tags": ["invoices"],
        "summary": "Update invoice",
        "operationId": "updateInvoice",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "name": "invoice_id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/InvoiceUpdate" } } } },
        "responses": {
          "200": { "description": "Invoice updated", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Invoice" } } } }
        }
      },
      "delete": {
        "tags": ["invoices"],
        "summary": "Delete invoice",
        "operationId": "deleteInvoice",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "name": "invoice_id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": { "204": { "description": "Invoice deleted" } }
      }
    },
...
    "/api/v1/ingestion/upload": {
      "post": {
        "tags": ["ingestion"],
        "summary": "Upload an invoice file for processing",
        "operationId": "uploadInvoice",
        "security": [{ "BearerAuth": [] }],
        "requestBody": { "required": true, "content": { "multipart/form-data": { "schema": { "type": "object", "properties": { "file": { "type": "string", "format": "binary", "description": "Invoice file (PDF, XML, PNG, JPEG, TIFF)" } }, "required": ["file"] } } } },
        "responses": {
          "200": { "description": "File uploaded and queued", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/FileUploadResponse" } } } }
        }
      }
    },
    "/api/v1/ingestion/webhook": {
      "post": {
        "tags": ["ingestion"],
        "summary": "Receive invoice via webhook",
        "operationId": "webhookReceive",
        "security": [{ "BearerAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/WebhookPayload" } } } },
        "responses": {
          "200": { "description": "Webhook received", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/FileUploadResponse" } } } }
        }
      }
    },
...
  },
  "security": [{ "BearerAuth": [] }]
}